+++ src/ConstraintProgrammingExtensions.jl
@@ -33,4 +33,7 @@ include("Test/Test.jl")
 include("FlatZinc/FlatZinc.jl")
 include("XCSP/XCSP.jl")
 
+# Compatibility layer.
+@deprecate Reified Reification ex=false
+
 end