+++ src/Test/indicator.jl
@@ -1,0 +1,82 @@
+function indicator_singlevariable_test(
+    model::MOI.ModelLike,
+    config::MOIT.TestConfig,
+)
+    MOI.empty!(model)
+
+    @test MOI.supports_constraint(model, MOI.SingleVariable, MOI.Integer)
+    @test MOI.supports_constraint(model, MOI.SingleVariable, MOI.EqualTo{Int})
+    @test MOI.supports_constraint(
+        model,
+        MOI.VectorOfVariables,
+        MOI.IndicatorSet{MOI.ACTIVATE_ON_ONE, MOI.EqualTo{Int}},
+    )
+
+    x1, _ = MOI.add_constrained_variable(model, MOI.Integer())
+    x2, _ = MOI.add_constrained_variable(model, MOI.Integer())
+
+    c1 = MOI.add_constraint(model, x1, MOI.EqualTo(1))
+    c2 = MOI.add_constraint(model, _vov([x1, x2]), MOI.IndicatorSet{MOI.ACTIVATE_ON_ONE}(MOI.EqualTo(2)))
+
+    @test MOI.is_valid(model, x1)
+    @test MOI.is_valid(model, x2)
+    @test MOI.is_valid(model, c1)
+    @test MOI.is_valid(model, c2)
+
+    if config.solve
+        MOI.optimize!(model)
+        @test MOI.get(model, MOI.TerminationStatus()) == MOI.OPTIMAL
+        @test MOI.get(model, MOI.PrimalStatus()) == MOI.FEASIBLE_POINT
+
+        @test MOI.get(model, MOI.ResultCount()) >= 1
+        @test MOI.get(model, MOI.VariablePrimal(), x1) == 1
+        @test MOI.get(model, MOI.VariablePrimal(), x2) == 2
+    end
+end
+
+function indicator_scalaraffinefunction_test(
+    model::MOI.ModelLike,
+    config::MOIT.TestConfig,
+)
+    MOI.empty!(model)
+
+    @test MOI.supports_constraint(model, MOI.SingleVariable, MOI.Integer)
+    @test MOI.supports_constraint(
+        model,
+        MOI.ScalarAffineFunction{Int},
+        MOI.EqualTo{Int},
+    )
+    @test MOI.supports_constraint(
+        model,
+        MOI.VectorAffineFunction{Int},
+        MOI.IndicatorSet{MOI.ACTIVATE_ON_ONE, MOI.EqualTo{Int}},
+    )
+
+    x1, _ = MOI.add_constrained_variable(model, MOI.Integer())
+    x2, _ = MOI.add_constrained_variable(model, MOI.Integer())
+
+    c1 = MOI.add_constraint(model, _saf(x1), MOI.EqualTo(1))
+    c2 = MOI.add_constraint(model, _vaf([x1, x2]), MOI.IndicatorSet{MOI.ACTIVATE_ON_ONE}(MOI.EqualTo(2)))
+
+    @test MOI.is_valid(model, x1)
+    @test MOI.is_valid(model, x2)
+    @test MOI.is_valid(model, c1)
+    @test MOI.is_valid(model, c2)
+
+    if config.solve
+        MOI.optimize!(model)
+        @test MOI.get(model, MOI.TerminationStatus()) == MOI.OPTIMAL
+        @test MOI.get(model, MOI.PrimalStatus()) == MOI.FEASIBLE_POINT
+
+        @test MOI.get(model, MOI.ResultCount()) >= 1
+        @test MOI.get(model, MOI.VariablePrimal(), x1) == 1
+        @test MOI.get(model, MOI.VariablePrimal(), x2) == 2
+    end
+end
+
+const indicatortests = Dict(
+    "indicator_singlevariable" => indicator_singlevariable_test,
+    "indicator_scalaraffinefunction" => indicator_scalaraffinefunction_test,
+)
+
+MOIT.@moitestset indicator