+++ NEWS.md
@@ -10,6 +10,7 @@ Generalise `Count` to any kind of comparison for items, not just equality.
 Add `VectorDomain` and `VectorAntiDomain`. Add `AllDifferentExceptConstants`;
 `AllDifferentExceptConstant` becomes a constructor for the former. Add 
 `NonOverlappingOrthotopes` and `ConditionallyNonOverlappingOrthotopes`.
+Add `GlobalCardinality` and `GlobalCardinalityVariable`.
 
 
 Version 0.2.5